fix: installer defaults to PuTTY
The installer would always default to PuTTY as the SSH client.
Whilst in an interactive mode a user is able to choose the SSH client,
a silent install (e.g. the auto-update) will reset the setting to PuTTY
despite users using OpenSSH previously.
This has caused issues like #7543 and numerous questions on SO.

The installer is updated to read the currently configured value from
the registry. If the value is absent (new installation) or unset - use
OpenSSH. Else presume PuTTY.

This leave users with custom configurations out in a cold (e.g. #6286),
but our telemetry (however limited) shows a very small number of those
users. And the effort to update the installer to cater for those users
does not justify the gain.